refactor(core): change ToRatio return type to float for consistency
- Updated `ToRatio` method in `Percent` struct to return `float` instead of `double`.
  - Ensures consistency with arithmetic operations and other related values in `Percent`.
  - Improves performance by avoiding unnecessary precision.

No breaking changes; all usages remain functionally compatible without adjustments.

+++ src/Bezoro.Core/Common/Primitives/Percent.cs
@@ -81,5 +81,5 @@ public readonly struct Percent : IEquatable<Percent>, IComparable<Percent>
 	/// <summary>
 	///     Converts the percentage to a decimal ratio (0.0–1.0).
 	/// </summary>
-	public double ToRatio() => Value / 100.0;
+	public float ToRatio() => Value / 100f;
 }
